backend/requirements.txt
aiofiles==0.6.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
MarkupSafe==1.1.1
protobuf==3.15.5
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
validate-docbr==1.8.2

PyJWT~=2.0.1